third_party/WebKit/LayoutTests/vr/latest/exclusive_requestFrame_called.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/vr_service.mojom.js"></script>
<script src="../resources/mock-vr-service.js"></script>
<canvas id="webgl-canvas"></canvas>
<script src="../resources/presentation-setup.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_session_test( (t, session) => {
  // Session must have a baseLayer or frame requests will be ignored.
  session.baseLayer = new VRWebGLLayer(session, gl);

  function onFrame(vrFrame) {
    t.step( () => {
      assert_true(vrFrame instanceof VRPresentationFrame);
    });
    t.done();
  }

  session.requestFrame(onFrame);
}, fakeDisplays["Pixel"], { exclusive: true },
"VRSession requestFrame properly calls the provided callback");

</script>
